feat: refactor api jornadas y camiones
