+++ apps/mfe-dashboard/src/pages/RegistroNuevaJornada.tsx
@@ -1,7 +1,7 @@
 import { useNavigate, NavLink } from "react-router-dom";
 import { useState, useEffect } from "react";
 import "./Dashboard.css";
-import { getJornadas } from "../api/jornadas";
+import { getJornadas } from "../../../../packages/api-client/src/services/jornadas";
 
 const menuItems = [
   { label: "Dashboard", path: "/dashboard", icon: "📊" },
