+++ apps/mfe-dashboard/tsconfig.app.json
@@ -24,5 +24,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"]
+  "include": ["src", "../../packages/api-client/src/services/camiones.ts", "../../packages/api-client/src/services/jornadas.ts"]
 }
